requirements.txt
# Core dependencies
PyYAML>=6.0
python-telegram-bot>=20.0
requests>=2.28.0
schedule>=1.2.0

# Development and building
pyinstaller>=5.0
setuptools>=65.0

# Optional (for enhanced features)
# cryptography>=41.0.0  # For encryption
# psutil>=5.9.0        # For system monitoring
# tabulate>=0.9.0      # For formatted tables